Moved some classes and structs to own files and Battle namespace.
